[VM Agent] Build antrea binaries with static link
This change is to resolve an issue that antrea-agent may not work on a VM
without the linked the libraries if the runtime OS is different from the
environment in which they were compiled. The solution is to disable CGO when
compiling antrea-agent and antctl for VM scenarios. And a dedicated item is
"vm-bin" introudced to build VM agent scenarios, which only build antrea-agent
and antctl bits.

Signed-off-by: wenyingd <wenyingd@vmware.com>