Fix race condition in pkg/apiserver/certificate unit tests
There was some "interference" between TestSelfSignedCertProviderRotate
and TestSelfSignedCertProviderRun. The root cause is that the
certutil.GenerateSelfSignedCertKey does not support a custom clock
implementation and always calls time.Now() to determine the current
time. It then adds a year to the current time to set the expiration time
of the certificate. This means that when rotateSelfSignedCertificate()
is called as part of TestSelfSignedCertProviderRotate, the new
certificate is already expired, and rotateSelfSignedCertificate() will
be called immediately a second time. By this time however,
TestSelfSignedCertProviderRotate has already exited, and we are already
running the next test, TestSelfSignedCertProviderRun. This creates a
race condition because the next test will overwrite
generateSelfSignedCertKey with a mock version, right as it is called by
the second call to rotateSelfSignedCertificate() from the previous
test's provider.

To avoid this race condition, we use a WaitGroup to ensure that the
provider is no longer running before moving on to the next test.

An alternative solution I considered was locking p.mutex before
overwriting generateSelfSignedCertKey in
mockGenerateSelfSignedCertKey. This would have technically avoided the
race. However, I think the selected solution is more correct, as
multiple tests calling the mock generateSelfSignedCertKey concurrently
could potentially have unintended consequences, even if we make the
assignment thread-safe.

Fixes #5977

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>