[e2e tests] Remove explicit sleep command for toolbox containers
An explicit sleep is not required as the default command in the
container is already "pause".
The Windows version of the toolbox image will not start when using
"sleep", as the command is not available.
Even on Linux, pause is superior to sleep for signal handling.

We also fix the "don't fragment" ping option for Windows.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>