[e2e tests] Wait for secondary IPs before running ping test (#6041)
For VLAN Secondary Network e2e tests, we need to wait for secondary IPs
to be available. Interface creation can take up to a few seconds after
the Pod is in the Running state, as it happens asynchronously. Otherwise
we can get the following error:

secondary_network_test.go:194: Error when pinging between interfaces: interface eth1 not found on vlan-pod2. err: <nil>

Fixes #6023

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>