Stop using projects.registry.vmware.com for user-facing images
We stop referencing the VMware Harbor registry
(projects.registry.vmware.com) in the Helm charts and K8s manifests that
we release to users, with the expectation that images for future
releases will not be available from that registry, as it is scheduled to
be EOL'd and we won't be mirroring them there.

Because the antrea DockerHub organization is part of the "Sponsored OSS"
program, there should be no concern about users running into issues
because of DockerHub rate-limits. We could also consider mirroring these
images to ghcr.io in the future, to offer users an alternative.

At the moment, we do not update references to other images in the
projects.registry.vmware.com/antrea project (images used for testing),
with the expectation that they will remain available for the foreseeable
future. In other words, while we will no longer be publishing new images
to projects.registry.vmware.com, we expect that existing images will
remain available through the same URLs.

For #6003

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>