Fix nil pointer dereference when ClusterGroup/Group is used
When an AppliedToGroup or AddressGroup is derived from a ClusterGroup or
Group, we used whether the existence of the source Group in the internal
group storage as the indicator of the type of AppliedToGroup or
AddressGroup. After the source Group is deleted, the AppliedToGroup or
AddressGroup was considered as a Group with its own selector mistakenly.
Accessing its selector would panic due to nil pointer dereference.

This patch makes the type of AppliedToGroup and AddressGroup more
explicit by adding a field "SourceGroup" to indicate it. If the source
Group can't be found in the storage, we just return nil to indicate the
Group selects nothing at the moment.

Signed-off-by: Quan Tian <qtian@vmware.com>