Fix ci/test-e2e-kind.sh script on macOS
The script was failing with:
```
Deleting VLAN subnets
Option "-br" is unknown, try "ip help".
```
when cleaning-up a test cluster.

This is because the ip command on macOS only supports a subset of the
command-line options supported on Linux. In any case, on macOS, the
command needs be wrapped with docker_run_with_host_net, since we are
looking for VLAN interfaces running in the Docker Linux VM.

As a minor improvement, we also add the "antrea" prefix to the external
server container we create, along with a short random suffix ($RANDOM is
sufficient for this) to avoid possible conflicts.

Fixes #6022

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>