[Multicluster] Fix cache options for controller Manager
A few issues were introduced by #5843 because of changes in the
sigs.k8s.io/controller-runtime interface.

The biggest issue was that the call to ctrl.NewManager was not using the
Options object populated earlier in the setupManagerAndCertController
function. Instead it was creating and using a new, incomplete Options
object.

Fixes #6149

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>