Fix a flaky failure in VM Agent e2e test "testExternalNode"
The failure occurs when the check is performed before the default route on the
uplink is removed, then two interfaces are found, which leads to a mismatch with
the expectation.

The fix is to introduce a retry when checking the VM network configurations, and
and a check on the count of the default route before return.

Signed-off-by: Wenying Dong <wenyingd@vmware.com>