Fall back to lenient decoding when strict decoding config fails
When upgrading Antrea, the config is usually updated first and there may
be some new fields added to it. If a yet update agent crashes during the
process, it will fail to restart due to the unknown new fields because
we do strict decoding only.

The patch adds lenient decoding when strict decoding fails to tolerate
unknown fields and duplicate fields. Other errors like malformed data
and unmatched type are still fatal.

Signed-off-by: Quan Tian <quan.tian@broadcom.com>