[Multicluster] Fix cache options for controller Manager (#6150)
A few issues were introduced by #5843 because of changes in the
sigs.k8s.io/controller-runtime interface.

The biggest issue was that the call to ctrl.NewManager was not using the
Options object populated earlier in the setupManagerAndCertController
function. Instead it was creating and using a new, incomplete Options
object.

Additionally, the decoder is no longer injected automatically, it needs to be
instantiated by us. Otherwise the admission webhook panics.
See https://github.com/kubernetes-sigs/controller-runtime/issues/2695

Fixes #6149

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>