Optimize Docker build for Multicluster image
We use Docker cache mounts to accelerate Docker builds with Go.
With cache mounts, we can cache package dependencies and the Go build
cache in a persistent fashion (i.e., persists across builds).

This is similar to #5918, which implemented the same change for the
"main" Antrea images.

With this change, we go from 5 minutes down to 30 seconds when
re-building the antrea/antrea-mc-controller image after a minor change
to the source code.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>