Fix possible race condition in SelfSignedCertProvider tests (#6166)
I have seen some cases in Github CI where it took slightly over 2
seconds for the Secret update to be processed in
TestSelfSignedCertProviderRotate. The test ended up failing.

The most likely cause is the well-known race condition between
List and Watch for informers when using the fake client. If the Secret
Update happens concurrently with the Informer starting, right
between the List and the Watch operations, we may lose the Update
event. We call WaitForCacheSync to avoid race conditions.

To be more conservative, we also increase the wait duration for
assertions to 5 seconds.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>