Fix action field in NetworkPolicyEvaluation for KNP
When processing Kubernetes NetworkPolicy, a default action
Allow is assigned in Antrea controller. This was not handled
corretly in NetworkPolicyEvaluation action field, for default
isolation model. As it will undeterminately display action.

This PR fixes the issue, by indicating the rule as default
isolation with nil action field, and then assigning Isolate
to the response in NetworkPolicyEvaluation.

Signed-off-by: Qiyue Yao <yaoq@vmware.com>