Fix VM Agent e2e test
The test was failing because of a missing expected entry for the
supportbundle (OVS groups, which was added recently).

Because this is not the first time that it happens, we relax the test so
that it no longer fails when the supportbundle actual contents are not
an exact match with the expected contents (they have to be a superset).

Fixes #6215

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>