Request basic memory for antrea-controller
Starting with K8s v1.22, system-cluster-critical Pods no longer get low
OOM score like system-node-critical Pods. antrea-controller's OOM score
adjustment was changed from -997 to 999 passively, which made it just
have a bit higher chance of survival over best-effort Pods (OOM score
adjustment: 1000). While it's hard to say how much memory it can use as
it's highly related to the scale of the cluster, we can add a basic
memory request which can decrease the score a bit according to the
formula for burstable Pods:

  oomScoreAdjust := 1000 - (1000*memoryRequest)/memoryCapacity

It should also be a bit helpful to avoid antrea-controller to be
scheduled on a Node with very few available memory.

Signed-off-by: Quan Tian <quan.tian@broadcom.com>