Remove incorrect AntreaProxy warning on Windows (#6242)
The code was logging an incorrect warning about AntreaProxy being
disabled. This is is because checkUnsupportedFeatures was in charge of
the check, but o.enableAntreaProxy is set later in the validation
chain. To avoid the issue, we introduce a new function for
platform-specific checks, validateConfigForPlatform, which runs after
all other validations and after all fields in the Options struct have
been set.

We also replace the warning message with an error message (but we do not
fail Agent initialization) and we add a check for proxyAll.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>