Fix that local traffic cannot be identified in networkPolicyOnly mode
Before this commit, in AntreaProxy, to respect short-circuiting, when
installing flows for an external Services, an extra flow with higher priority
to match traffic sourced from local (local Pods or local Node) and destined
for the external Service will be installed. This is achieved by matching
the local Pod CIDR obtained from the local Node object. However, when
Antrea is deployed in networkPolicyOnly mode, the Pod CIDR in the local Node
object is nil, resulting in the failure of install the extra flow mentioned
above. To fix the issue, a new reg mark `FromLocalRegMark` identifying traffic
from local Pods or the local Node is introduced to mark the traffic from
local. This reg mark can be used in all traffic mode.

Fix #6244

Signed-off-by: Hongliang Liu <lhongliang@vmware.com>