[Windows] Finish removing Docker and userspace kube-proxy support (#6255)
* Finish removing Docker and userspace kube-proxy support

We update the Windows documentation to remove all Docker-specific
instructions, and all mentions of (userspace) kube-proxy.

We also update Windows scripts to remove Docker support. The
Prepare-AntreaAgent.ps1 script was still defaulting to installing
kube-proxy, so we update the script to change this default behavior. For
all the scripts, "omitting" kube-proxy is now the default behavior. The
corresponding script parameters have not been removed yet, but they are
not officially deprecated, and we can remove them in a future release,
such as Antrea v2.3.

Fixes #5630

* Rename Antrea Windows YAML manifests

Now that containerd is the only supported container runtime, we rename
antrea-windows-containerd.yml to antrea-windows.yml and
antrea-windows-containerd-with-ovs.yml to antrea-windows-with-ovs.yml.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>