Minor fix to start_ovs script when bridging mode is enabled
The SLEEP_PID variable (used to store the PID of the sleep process
started by the "main" loop) was being overwritten in the signal handler.
The 1-second sleep which runs in the signal handler and wait for uplink
port deletion (when bridging mode is enabled) does not need to run in
the background.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>