Set up Docker Buildx in workflows whenever building Antrea (#6320)
Whenever the Antrea images are built from scratch using
`./hack/build-antrea-linux-all.sh` in Github CI, we should set up Docker
Buildx first. It ensures that the docker-container driver is used, which
enables the registry cache backend. We depend on this cache backend to
speed up the build by avoiding re-building the base images if it is not
necessary. Without caching, the build takes ~10 minutes instead of ~3
minutes.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>