Fix trivy Github workflow
Ever since splitting up the Agent and Controller images, there was an
issue with the Trivy workflow, as the same output path was used for the
Trivy report for both the Agent and Controller. When generating the
report for the Controller, it would overwrite the one for the Agent,
meaning that incomplete information was uploaded as job artifacts.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>