Fix RestoreHostInterfaceConfiguration for SecondaryNetwork
When SecondaryNetwork is enabled, but the ovsBridges list is empty, the
antrea-agent currently panics during graceful termination. This is
because the RestoreHostInterfaceConfiguration was assuming that there
was alwasy one OVS bridge.

The panic is not really impactful, as it happens during Pod
termination. However, for code coverage collection in e2e tests, it can
have an impact, so we want to make sure it doesn't happen.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>