Support simple ping mesh in Agent (#6120)
We introduce a new feature to measure inter-Node latency in a K8s
cluster running Antrea. The feature is currently Alpha and uses the
NodeLatencyMonitor FeatureGate.

In addition to the FeatureGate, enablement of the feature is controlled
by a new CRD, called NodeLatencyMonitor. This CRD supports at most one
CR instance, which must be named "default". When the CR exists, Antrea
Agents will start "pinging" each other to take latency measurements.

Each Agent only stores the latest measured value (at least at the
moment), we do not store time series data.

We support both IPv4 and IPv6. When an oberlay is used by Antrea, the
ping is sent over the tunnel (by using the gateway IP as the
destination).

This change does not add any functionality besides collecting latency
data at each Agent. A follow-up change will take care of reporting the
latency data to the Antrea Controller, so it can be consumed via an
APIService.

For #5514

Signed-off-by: IRONICBo <boironic@gmail.com>
Signed-off-by: Asklv <boironic@gmail.com>