[e2e tests] Remove custom timeout value for coverage image
There should no longer be a need for a custom timeout value in
testAntreaGracefulExit when coverage is enabled. The increased timeout
was not actually required for code coverage collection (which is qite
fast), but because of issues with the coverage-enabled image, which was
not handling signals correctly. After #6090, this should have been
fixed.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>