Fix hack/make-metrics-doc.sh script (#6393)
From time to time, the script will fail in CI with the following error:
```
error: no matching resources found
```

This error occurs when the kubectl command used to wait for all Antrea
Pods to become Ready is run but no Pod has been created yet. This is
possible, especially on a slow machine, as applying the Antrea manifest
does not create any Pod. Instead, it creates the antrea-agent DaemonSet
and the antrea-controller Deployment, and the K8s control-plane may take
some time to actually create the corresponding Pods. To fix this issue,
we first wait for the rollout to complete for both antrea-agent and
antrea-controller.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>