Always include Pod labels in FlowAggregator IPFIX template
We always include the Pod labels IEs (for source and destination Pods),
regardless of the value of the recordContents.podLabels configuration
parameter. This simplifies the logic and the IPFIXExporter no longer
needs to be aware of this configuration. There will be a minor size
increase to the IPFIX records exported by the FlowAggregator when
recordContents.podLabels is false, as we will need to include empty
strings in the records for the 2 IEs.

We use an empty string when recordContents.podLabels is false, or when
the endpoint is not a Pod. We use an empty JSON dictionary ("{}"), when
the Pod has no labels.

Fixes #6386

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>