Handle unexpected error when the test is aborted (#5295)
When the conformance test is aborted by Jenkins, the results log
is actually an incomplete file. Since run-k8s-e2e-tests.sh already
checks the result and returns with different exit error codes, we can
check the codes and print the right message. Otherwise, the result will show
`All tests passed.` when the test is actually running and aborted
unexpectedly, which is misleading.

Signed-off-by: Lan Luo <luola@vmware.com>