Fix racy unit test TestIdAllocatorWorker (#6407)
The test has been failing in Github CI, albeit very rarely.
The issue is that there exists a possible race between the creation of
the timer by the DelayingQueue (which determines when to add the item to
the queue after the waiting period is over) and the test code that
advances the clock. If the DelayingQueue determines that a 1s timer is
appropriate, but we advance the fake clock right as NewTimer is being
called, we will actually sleep until a deadline that is further in the
future than expected. The K8s FakeClock implementation provides an
HasWaiters method which is meant to be used by tests to avoid such
races. However, it is not sufficient in our case, since the creation of
the DelayingQueue will instantiate a Ticker, which itself counts as a
waiter. Therefore HasWaiters cannot be used as the signal indicating
that the Timer has been created. To address this limitation, we wrap
FakeClock into a separate struct that tracks the number of times
NewTimer has been called.

An alternative solution would have been to call AddAfter on the
DelayingQueue (with an arbitrary item and a duration well into the
future) after advancing the clock, as any call to AddAfter will trigger
the processing of waiting items. However, this is not a great solution
IMO as it relies too much on knowledge of the implementation internals
of DelayingQueue.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>