antctl: improve deny-all NetworkPolicy check (#6412)
The egress-deny-all-connectivity check takes more than 1m to run even if
it works as expected. This is because the client Pod can't resolve the
Service name to IP due to the deny-all policy while DNS resolution
doesn't have an explicit timeout.

The inability to resolve Service names make testing both SameNodeService
and OtherNodeService pointless as the client won't really send any
traffic to them.

The commit changes to test ClusterIPs of target Services directly, to
check the NetworkPolicy enforcement of both inter-Node and intra-Node
traffic, and both IPv4 and IPv6 family. It also reduces the code
redundancy between egress-deny-all and ingress-deny-all tests.

Signed-off-by: Quan Tian <quan.tian@broadcom.com>