Download OpenSSL from Github for Windows Docker build
Indy’s Fulgan mirror is currently down, preventing the image from
building. This is not the first time apparently, and the OpenSSL
binaries are also available in Indy's Github repository.
See https://www.indyproject.org/2020/06/16/openssl-binaries-moved-to-github/

In the future, we could even consider keeping a copy of the archive in
downloads.antrea.io if needed.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>