Use create reactor for NodeLatencyStats fake client (#6492)
The monitor uses create to report the latest latency measurements, and
we do not need to distinguish between create and update for our use
case. However, the fake client is implemented using a simple object
tracker that will fail if create is called multiple times for the same
resource name. This can affect unit tests for the monitor, either
because a misleading error message is logged or because updates are
ignored.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>