Fix codecov upload for Jenkins CI scripts (#6493)
It seems that codecov upload has been broken for a while for
ci/jenkins/test-vmc.sh and ci/jenkins/test-mc.sh. The file pattern must
be quoted when calling the codecov binary, or the glob will be expanded
by the shell instead of by codecov (which is bad because we want the
glob to be expanded in the context of the search directory).

We also set "disable_search" to "true" when using the codecov-action in
Github workflows. This action uses the codecov CLI and there is no need
to keep search enabled when providing an explicit and comprehensive list
of files to include in the report.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>